% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\docType{class}
\name{Decoder-class}
\alias{Decoder-class}
\title{Decoder S4 Class}
\description{
An S4 class to represent a neural network decoder, used for emulating spatial priors.
The class includes parameters for input and output weight matrices and biases, as well as region identifiers.
}
\section{Slots}{

\describe{
\item{\code{GEOID}}{A character vector of region or area identifiers.}

\item{\code{W_in}}{An array representing the input weight matrix of the decoder.}

\item{\code{B_in}}{An array representing the input bias vector of the decoder.}

\item{\code{W_out}}{An array representing the output weight matrix of the decoder.}

\item{\code{B_out}}{An array representing the output bias vector of the decoder.}
}}

