% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.r
\name{load_environment}
\alias{load_environment}
\title{Load the vgmsfh numpyro module.}
\usage{
load_environment(envname = "vmsae")
}
\arguments{
\item{envname}{String, string path to python virtual environment.}
}
\description{
This function loads the vgmsfh numpyro module
}
