% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{load_vae}
\alias{load_vae}
\title{Load pretrained VAE decoder.}
\usage{
load_vae(model_name, save_dir = NULL)
}
\arguments{
\item{model_name, }{String, trained VAE model name in lower case.}

\item{save_dir, }{String, directory to save the trained VAE model. Default to the current directory.}
}
\value{
An object of the Decoder class.
}
\description{
Load pretrained VAE decoder.
}
