% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh_utils.r
\name{plot,VGMSFH-method}
\alias{plot,VGMSFH-method}
\title{Plot VGMSFH result.}
\usage{
\S4method{plot}{VGMSFH}(x, shp = NULL, var_idx = 1, type = "compare")
}
\arguments{
\item{shp}{The shapefile that contains the geometry. If null, this is downloaded from the pretrain model.}

\item{var_idx}{Index of the variable of interest.}

\item{type}{Type of plots. "compare" is the plot comparing direct estimate and model estimate. "estimate" is the plot that shows the mean and standard deviation of the model estimate.}

\item{object}{An object containing the VGMSFH results.}
}
\description{
This function plot the spatial map of VGMSFH result.
}
