% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh_utils.r
\name{plot.VGMSFH}
\alias{plot.VGMSFH}
\title{Plot VGMSFH result.}
\usage{
\method{plot}{VGMSFH}(object, shp = NULL, var_idx = 1, field = "yhat_samples", ...)
}
\arguments{
\item{object}{An object containing the VGMSFH results.}

\item{var_idx}{Index of the variable of interest.}

\item{field}{The field from the object to plot.}
}
\description{
This function plot the spatial map of VGMSFH result.
}
