% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh_utils.r
\name{summary,VGMSFH-method}
\alias{summary,VGMSFH-method}
\title{Summarize VGMSFH result.}
\usage{
\S4method{summary}{VGMSFH}(object, var_idx = 1, field = "beta_samples")
}
\value{
data frame, summary of VGMSFH result.
}
\description{
This function display the summary of VGMSFH result.
}
