% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh.r
\name{summary.VGMSFH}
\alias{summary.VGMSFH}
\title{Summarize VGMSFH result.}
\usage{
\method{summary}{VGMSFH}(object, var_idx = 1, field = "beta")
}
\value{
data frame, summary of VGMSFH result.
}
\description{
This function display the summary of VGMSFH result.
}
