% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{train_vae}
\alias{train_vae}
\title{Train VAE for spatial random effect.}
\usage{
train_vae(
  W,
  GEOID,
  vae_model_name,
  vae_save_dir = ".",
  n_samples,
  batch_size,
  epoch,
  lr_init = 0.001,
  lr_min = 1e-07
)
}
\arguments{
\item{W}{Matrix, proximity matrix.}

\item{n_samples, }{Int, number of training samples drawn from the prior.}

\item{batch_size, }{Int, batch size of VAE.}

\item{epoch, }{Int epoch of VAE.}

\item{lr_init, }{Float, initial learning rate.}

\item{lr_min, }{Float, reduced learning rate at the last epoch.}

\item{save_path, }{String, path to save the trained VAE model.}
}
\description{
This function trains a VAE that learn the conditional autoregressive prior.
}
