% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh.r
\docType{class}
\name{VGMSFH-class}
\alias{VGMSFH-class}
\title{Class contains the results of VGMSFH.}
\description{
Class contains the results of VGMSFH.
}
\section{Slots}{

\describe{
\item{\code{direct_estimate}}{array, Direct estimate of the population parameters.}

\item{\code{yhat_samples}}{array, The posterior samples of estimated population parameters.}

\item{\code{spatial_samples}}{array, The posterior samples of estimated spatial random effects.}

\item{\code{beta_samples}}{array, The posterior samples of coefficients of fixed effects.}

\item{\code{all_samples}}{array, The posterior samples of all parameters in VGMSFH.}
}}

