% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh.r
\name{vgmsfh_stan}
\alias{vgmsfh_stan}
\title{Run vgmsfh using stan.}
\usage{
vgmsfh_stan(
  y,
  y_sigma,
  X,
  W,
  GEOID,
  vae_model_name,
  vae_save_dir = NULL,
  num_samples = 1000,
  num_warmup = 1000,
  chains = 1,
  thin = 1
)
}
\arguments{
\item{y}{Matrix, responses.}

\item{y_sigma}{Matrix, reported standard deviations.}

\item{X}{Matrix, covariance matrix.}

\item{W}{Matrix, proximity matrix}

\item{vae_model_name}{String, vae model name.}

\item{vae_save_dir}{String, vae model saving directory.}

\item{num_samples}{Int, Number of posterior samples.}

\item{num_warmup}{Int, Number of burning-in.}
}
\value{
Matrix, the true process.
}
\description{
This function runs vgmsfh using stan as the backend.
}
