% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\docType{class}
\name{Decoder-class}
\alias{Decoder-class}
\title{Class representing a VAE decoder.}
\description{
Class representing a VAE decoder.
}
\section{Slots}{

\describe{
\item{\code{GEOID}}{character vector, The FIPS codes or other equivalent GEOIDs.}

\item{\code{W_in}}{array, The input weights.}

\item{\code{B_in}}{array, The input bias.}

\item{\code{W_out}}{array, The output weights.}

\item{\code{B_in}}{array, The output bias.}
}}

