% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.r
\name{download_pretrained_vae}
\alias{download_pretrained_vae}
\title{Download pretrained VAE model.}
\usage{
download_pretrained_vae(vae_model_name, save_dir)
}
\arguments{
\item{vae_model_name}{String, vae model name. e.g. "mo_county".}
}
\description{
This function downloads pretrained VAE model and the corresponding GEOID.
}
