% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.r
\name{install_environment}
\alias{install_environment}
\title{create the python environment.}
\usage{
install_environment(envname = "vmsae")
}
\arguments{
\item{envname}{String, string path to python virtual environment.}
}
\description{
This function loads the vgmsfh numpyro module
}
