% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.r
\name{install_environment}
\alias{install_environment}
\title{Install python environment.}
\usage{
install_environment(envname = "vmsae", use_gpu = FALSE)
}
\arguments{
\item{envname}{Character. The name of the Python environment to create or update.
Default is \code{"vmsae"}.}

\item{use_gpu}{Boolean. An indicator for whether to install packages with GPU support.
Default is \code{FALSE}.}
}
\value{
No return value, called for side effects
}
\description{
This function creates the vmsae python environment and installs required packages.
}
\examples{
\dontrun{
library(vmsae)
# this function is time consuming for the first run
install_environment()          # Install into default "vmsae" environment


# this step is time consuming for the first run
install_environment("custom")  # Install into a custom-named environment
}

}
