% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh.r
\docType{class}
\name{VGMSFH-class}
\alias{VGMSFH-class}
\title{VGMSFH S4 Class}
\description{
An S4 class to store results from the Variational Gaussian Markov Small Area Estimation with Fay-Herriot model (VGMSFH). This class holds the posterior samples for various model components as well as the original direct estimates.
}
\section{Slots}{

\describe{
\item{\code{model_name}}{Character. The name of the trained VAE model.}

\item{\code{direct_estimate}}{Array. Direct estimates of parameters.}

\item{\code{yhat_samples}}{Array. Posterior samples of the estimated parameters.}

\item{\code{spatial_samples}}{Array. Posterior samples of the estimated spatial random effects.}

\item{\code{beta_samples}}{Array. Posterior samples of the fixed effect coefficients.}

\item{\code{all_samples}}{List. Posterior samples of all parameters in the VGMSFH model.}
}}

