% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgmsfh.r
\name{vgmsfh_numpyro}
\alias{vgmsfh_numpyro}
\title{Run vgmsfh using numpyro.}
\usage{
vgmsfh_numpyro(
  y,
  y_sigma,
  X,
  W,
  GEOID,
  vae_model_name,
  vae_save_dir = NULL,
  num_warmup = 1000,
  num_samples = 1000
)
}
\arguments{
\item{y}{Matrix, responses.}

\item{y_sigma}{Matrix, reported standard deviations.}

\item{X}{Matrix, covariance matrix.}

\item{W}{Matrix, proximity matrix}

\item{GEOID}{Vector, FIPS codes or other equivalent GEOIDs.}

\item{vae_model_name}{String, vae model name.}

\item{vae_save_dir}{String, vae model saving directory. Default to use pretrained models.}

\item{num_warmup}{Int, Number of burning-in. Default to 1000.}

\item{num_samples}{Int, Number of posterior samples. Default to 1000.}
}
\value{
VGMSFH s4 object, which contains the direct estimate and the posterior samples of yhat (population process), mu (intercept of population process), beta (coefficient of covariates), delta (fine scale variations of population process), car (spatial random effects of population process). In addition, all other posteriors (bridging variables, variations of random effects) including all meantioned before are provided in "all_samples".
}
\description{
Run vgmsfh using numpyro as the backend.
}
